+++ infra/aws_baseline_80/service_discovery.tf
@@ -126,3 +126,24 @@ resource "aws_service_discovery_service" "stt_http" {
 
   tags = local.tags
 }
+
+
+resource "aws_service_discovery_service" "llama_server" {
+  name = "llama-server"
+
+  dns_config {
+    namespace_id   = aws_service_discovery_private_dns_namespace.this.id
+    routing_policy = "MULTIVALUE"
+
+    dns_records {
+      type = "A"
+      ttl  = 10
+    }
+  }
+
+  health_check_custom_config {
+    failure_threshold = 1
+  }
+
+  tags = local.tags
+}